pyproject.toml
[project]
name = "lsp-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "debugpy>=1.8.14",
    "fastmcp>=2.5.1",
    "google-genai>=1.16.1",
    "lsprotocol>=2023.0.1",
    "pylspclient>=0.1.2",
    "python-lsp-server>=1.12.2",
    "ruff>=0.11.11",
]
